seperate Navigation into multiple components
