+++ src/components/containers/DefaultContainer.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Outlet } from "react-router-dom";
 import { Offline, Online } from "react-detect-offline";
 import Footer from "../shared/Footer";
-import Navigation from "../shared/Navigation";
 import OfflineContainer from "./OfflineContainer";
+import { Navigation } from "../navigation";
 
 const DefaultContainer = () => {
   return (
